gtav_selfdrive/main.py
import numpy as np
from PIL import ImageGrab
import cv2
import time
from numpy import ones,vstack
from numpy.linalg import lstsq
from libs.directkeys import PressKey, W, A, S, D, ReleaseKey, PressKey
from libs.draw_lanes import draw_lanes
from libs.image import grab_screen
from statistics import mean

def roi(img, vertices):
    
    #blank mask:
    mask = np.zeros_like(img)   
    
    #filling pixels inside the polygon defined by "vertices" with the fill color    
    cv2.fillPoly(mask, vertices, 255)
    
    #returning the image only where mask pixels are nonzero
    masked = cv2.bitwise_and(img, mask)
    return masked

def straight():
    PressKey(W)
    ReleaseKey(A)
    ReleaseKey(D)

def left():
    PressKey(A)
    ReleaseKey(W)
    ReleaseKey(D)
    ReleaseKey(A)

def right():
    PressKey(D)
    ReleaseKey(A)
    ReleaseKey(W)
    ReleaseKey(D)

def slow_ya_roll():
    ReleaseKey(W)
    ReleaseKey(A)
    ReleaseKey(D)

def main():
  last_time = time.time()
  while True:
    screen = grab_screen(region=(0,40,800,640))
    screen = cv2.cvtColor(screen, cv2.COLOR_BGR2GRAY)
    screen = cv2.resize(screen, (16,12))
    cv2.imwrite('screen.png',screen)
    print(screen[1][1])
    print(screen[2])
    print(screen[3])
    print('Frame took {} seconds'.format(time.time()-last_time))
    last_time = time.time()
    cv2.imshow('window', screen)

    #cv2.imshow('window',cv2.cvtColor(screen, cv2.COLOR_BGR2RGB))
    if cv2.waitKey(25) & 0xFF == ord('q'):
      cv2.destroyAllWindows()
      break
    break
if __name__== "__main__":
  main()