requirements.txt
tensorflow
pandas
tflearn
numpy
opencv-python
Pillow
pywin32
matplotlib